+++ CHANGELOG.md
@@ -3,6 +3,29 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.0.0][1.2.0.0]
+
+### Changed
+*   Refactored all examples.
+*   Fixed iDeal example.
+*   Metadata are no longer automatically created and sent when they are not set by the merchant.
+*   Version and Type of the SDK are now sent to the API with every request.
+*   Add missing feature to readme.
+*   Renamed property `Basket::amountTotal` to `Basket::amountTotalGross` to follow the change in the API.
+*   Refactor tests.
+*   Set `Paypage` default action to 'charge' and restrict values to 'charge' and 'authorize'.
+
+### Added
+*   Property `type` to `BasketItem`.
+*   An exception is now thrown in case of a `CURL` timeout.
+*   The `CURL` timeout can now be changed via environment variable.
+
+### Removed
+*   Unnecessary resource setters in `Paypage`.
+
+### Fixed
+*   Unit tests work now even in `development` or `staging` environment.
+
 ## [1.1.6.0][1.1.6.0]
 
 ### Added
@@ -219,3 +242,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [1.1.4.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.3.0..1.1.4.0
 [1.1.5.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.4.0..1.1.5.0
 [1.1.6.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.5.0..1.1.6.0
+[1.2.0.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.6.0..1.2.0.0